finn

+++ backend/AAA/AAA/Models/Orders.cs
@@ -17,7 +17,7 @@ namespace AAA.Models
         [Required]
         public string? imageUrl { get; set; }
         [Required]
-        public int counter { get; set; }
+        public long counter { get; set; }
 
     }
 }
